requirements.txt
blis==0.2.4
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
ChatterBot==1.0.5
cryptography==2.8
cymem==2.0.3
future==0.18.2
idna==2.8
mathparse==0.1.2
murmurhash==1.0.2
nltk==3.4.5
numpy==1.18.1
Pint==0.10.1
plac==0.9.6
preshed==2.0.1
pycparser==2.19
pymongo==3.10.1
python-dateutil==2.7.5
python-telegram-bot==12.3.0
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
six==1.14.0
spacy==2.1.9
SQLAlchemy==1.2.19
srsly==1.0.1
thinc==7.0.8
tornado==6.0.3
tqdm==4.42.1
urllib3==1.25.8
wasabi==0.6.0
